backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.1
pydantic==2.10.4
pydantic-settings==2.7.1
httpx==0.28.1
beautifulsoup4==4.12.3
anthropic==0.43.0
openai==1.58.1
google-genai==1.0.0
pillow==11.1.0
python-multipart==0.0.20
playwright==1.50.0
